Revert "fix:バックエンドの実装、フロントエンドの修正"
