+++ backend/src/env.ts
@@ -4,7 +4,7 @@ const envSchema = z.object({
   NODE_ENV: z.enum(["development", "production", "test"]).default("development"),
   PORT: z.coerce.number().default(3000),
   CORS_ORIGIN: z.string().default("https://minpuzz.vercel.app"),
-  REDIS_URL: z.string().default("redis://redis:6379"),
+  REDIS_URL: z.string().default("redis://localhost:6379"),
 });
 
 try {
