+++ backend/src/model/game/init.ts
@@ -13,7 +13,6 @@ export const GameInitPayloadSchema = z.object({
       placed: z.boolean(),
       row: z.number().int().nonnegative().optional(),
       col: z.number().int().nonnegative().optional(),
-      season: z.enum(['spring', 'summer', 'autumn', 'winter']).optional(),
     })
   ),
   startedAt: z.string().nullable(),
@@ -22,13 +21,13 @@ export const GameInitPayloadSchema = z.object({
 
 export type GameInitPayload = z.infer<typeof GameInitPayloadSchema>;
 
-// M1: 仕様に基づくダミー初期化データ（固定の5x5、ピースなし）
+// M1: 仕様に基づくダミー初期化データ（固定の6x6、ピースなし）
 export function buildInitPayload(params: { matchId: string; teamId: string; userId: string }): GameInitPayload {
   const payload: GameInitPayload = {
     matchId: params.matchId,
     teamId: params.teamId,
     userId: params.userId,
-    board: { rows: 5, cols: 5 },
+    board: { rows: 6, cols: 6 },
     pieces: [],
     startedAt: null,
     durationMs: null,
@@ -47,7 +46,7 @@ export function buildInitPayloadWithTimer(
     matchId: params.matchId,
     teamId: params.teamId,
     userId: params.userId,
-    board: { rows: 5, cols: 5 },
+    board: { rows: 6, cols: 6 },
     pieces: [],
     startedAt: timer ? timer.startedAt : null,
     durationMs: timer ? timer.durationMs : null,
